+++ src/nodes/navigation/GoToFolder/index.tsx
@@ -1,15 +1,24 @@
-import { Button, Node } from "@/components";
+import { Button, Node as BaseNode } from "@/components";
 import { Browsers } from "@/icons/Browsers";
-import { Handle, NodeProps, Position, useStore } from "reactflow";
+import {
+  Handle,
+  type Node,
+  type NodeProps,
+  Position,
+  useStore,
+} from "@xyflow/react";
 import { InvisibleArea, bottomHandleStyle, topHandleStyle } from "./styles";
 import { getMainFolder } from "./utils";
 import { useGoToFolder } from "./hooks";
 import { useOrchNode } from "@/nodes";
 
-interface IGoToFolderProps {
-  path?: string;
-  onClickFolder?: (nodeId: string) => void;
-}
+type IGoToFolderProps = Node<
+  {
+    path?: string;
+    onClickFolder?: (nodeId: string) => void;
+  },
+  "gotofolder"
+>;
 
 export const GoToFolder = ({
   data: { path, onClickFolder },
@@ -19,19 +28,19 @@ export const GoToFolder = ({
   const { isTarget } = useOrchNode({ nodeId: rest.id });
 
   return (
-    <Node.Root scheme="blue" {...rest}>
-      <Node.Title scheme="blue">
+    <BaseNode.Root scheme="blue" {...rest}>
+      <BaseNode.Title scheme="blue">
         <Browsers size={20} weight="regular" /> go to folder
         {path && <b>: {getMainFolder(path)}</b>}
-      </Node.Title>
-      <Node.Content>
+      </BaseNode.Title>
+      <BaseNode.Content>
         <Button
           onClick={() => onClickFolder?.(rest.id)}
           variant={path ? "link" : "ghost"}
         >
           {folderContent}
         </Button>
-      </Node.Content>
+      </BaseNode.Content>
       <Handle
         id="top"
         style={{
@@ -50,6 +59,6 @@ export const GoToFolder = ({
         position={Position.Bottom}
         type="source"
       />
-    </Node.Root>
+    </BaseNode.Root>
   );
 };