+++ src/stories/pages/nodes/GoToFolder.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { fn } from "@storybook/test";
 import { GoToFolder } from "@/nodes";
-import { ReactFlowProvider } from "reactflow";
+import { ReactFlowProvider } from "@xyflow/react";
 import { Center } from "../../components/Center";
 
 const meta: Meta = {